Remove ASM references from SandboxClassLoader.
